requirements.txt
numpy==1.24.4
hic-straw==1.3.1
networkx==2.6.3
karateclub==1.3.3
pandas==1.3.5
scipy==1.10.1
seaborn==0.13.2
matplotlib==3.8.4
scikit-learn==1.5.0
nx-cugraph-cu12==24.6.1